requirements.txt
cryptography>=42.0.0
Flask>=3.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
pdfminer.six>=20231228
Flask>=3.0.0
